% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.2nd}
\alias{cheb.2nd}
\title{Generating Chebyshev Polynomials of second kind}
\usage{
cheb.2nd(x.axis, n)
}
\arguments{
\item{x.axis}{beliebigie X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb.u Chebyshev-Polynome Zweiter Art (Vektor)
}
\description{
cheb.2nd erzeugt Chebyshev Polynome zweiter Art aus beliebiger X-Achse
}
\examples{
cheb.u <- cheb.2nd(x.axis, n)
}

