% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.fit.roots}
\alias{cheb.fit.roots}
\title{Curve Fitting with Chebyshev Polynomials and Finding of its Roots}
\usage{
cheb.fit.roots(d, x.axis, n, bc.harmonic = FALSE, roots.bound.l = NA,
  roots.bound.u = NA)
}
\arguments{
\item{d}{Zu fittender Datensatz/Zeitreihe (Vektor)}

\item{x.axis}{Beliebige X-Achse (Vektor)}

\item{n}{Ordnung des Polynoms (Skalar)}
}
\value{
cheb.list Berechnete Parameter (Koeffizienten, gefiltertes Modell, erste und zweite Ableitung des gefilterten Modells, Extremstellen und -Werte) (Liste)
}
\description{
\code{cheb.fit} fittet ein Chebyshev-Polynom beliebiger Ordnung an einen Datensatz/Zeitreihe mittels Least Squares Verfahren
}
\examples{
cheb.list <- cheb.fit(d, x.axis, n)
}

