% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.fit.seq}
\alias{cheb.fit.seq}
\title{Curve Fitting with Chebyshev Polynomials over Sequences}
\usage{
cheb.fit.seq(d, x.axis, n, l)
}
\description{
Fittet ein Chebyshev Polynom beliebiger Ordnung an einen sequenzierten Datensatz/Zeitreihe mittels Least Squares Verfahren
}
\examples{
cheb.fit.seq(d, x.axis, n, l)
}

