% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.model.filter}
\alias{cheb.model.filter}
\title{Calculation of Values of the model fit}
\usage{
cheb.model.filter(x.axis, cheb.coeff)
}
\arguments{
\item{x.axis}{beliebige X-Achse (Skalar oder Vektor)}

\item{cheb.coeff}{Chebyshev-Koeffizienten aus Least-Squares-Verfahren (Vektor)}
}
\value{
cheb.model gefiltertes Modell (Skalar oder Vektor)
}
\description{
cheb.model berechnet aus den Chebyshev-Koeffizienten die Y-Werte
}
\examples{
cheb.model <- cheb.model.filter(x.axis, cheb.coeff)
}

